refactor(discover): extract filter state into useDiscoverFilters hook
