+++ src/app/(user)/discover/_components/ExploreMapView.tsx
@@ -6,7 +6,7 @@ import { useToast } from "../../_hooks/useToast";
 import { dedupeEventMarkers } from "@/lib/event-utils";
 import { EVENT_RED, getDDay, sortEventMarkers } from "@/lib/event-format";
 import { useSearchParams, useRouter, usePathname } from "next/navigation";
-import { parseFilterParams, serializeFilterParams, buildTopicColorMap, KPOP_NAME, topicSlugToId } from "@/lib/filter-params";
+import { buildTopicColorMap } from "@/lib/filter-params";
 import { postMatchesFilters, placeMatchesFilters, placeMatchScore } from "@/lib/discover-filter-utils";
 import { InteractiveMap, type FocusCameraHandle } from "@/components/maps/InteractiveMap";
 import { PlaceBottomSheet } from "@/components/maps/PlaceBottomSheet";
@@ -24,15 +24,9 @@ import { HotTabStub } from "./HotTabStub";
 import { ScrollToTopButton } from "../../_components/ScrollToTopButton";
 import { useRecentSearches } from "../_hooks/useRecentSearches";
 import { useDiscoverViewState } from "../_hooks/useDiscoverViewState";
+import { useDiscoverFilters } from "../_hooks/useDiscoverFilters";
 import type { MapPlace, MapPost } from "@/lib/map-queries";
 import { getTopicMarkerColor, getTopicMarkerGradient, topicMatchesFilter, matchesQuery } from "@/lib/map-utils";
-import { getPlaceRegionSlug, getPlaceRegionLabel } from "@/lib/region-utils";
-import {
-  resolveTopicColors,
-  resolveTagColors,
-  labelBackground,
-  DEFAULT_TEXT,
-} from "@/lib/post-labels";
 import type { Level0TopicDeep } from "@/lib/topic-queries";
 import type { TagGroupWithTags } from "@/lib/filter-queries";
 import type { TagGroupColorMap } from "@/lib/post-labels";
@@ -43,7 +37,6 @@ import type {
 } from "@/lib/event-collection-queries";
 import type { CuratedSectionWithSlug, SectionData } from "@/lib/curation-types";
 
-type ChipInfo = { id: string; label: string; bg: string; fg: string };
 const CATEGORY_ORDER: string[] = [
   "CONCERT", "LANDMARK_LIGHTING", "PROMOTION", "ACTIVITY",
   "SHOPPING", "MOBILITY", "FNB", "STAY", "WELCOME_KIT",
@@ -98,38 +91,51 @@ export function ExploreMapView({ allPlaces, savedPostIds, savedEventIds = [], ta
   const [selectedCategory, setSelectedCategory] = useState<string | null>(null);
   const [savedOnly, setSavedOnly] = useState(false);
   const [isSearchOpen, setIsSearchOpen] = useState(false);
-  const [isFilterOpen, setIsFilterOpen] = useState(false);
-  const [stagedTopicIds, setStagedTopicIds] = useState<string[]>([]);
-  const [stagedTagIds, setStagedTagIds] = useState<string[]>([]);
-  const [stagedTagGroupKeys, setStagedTagGroupKeys] = useState<string[]>([]);
-  const [stagedRegion, setStagedRegion] = useState<string | null>(null);
-  const [appliedTopicIds, setAppliedTopicIds] = useState<string[]>([]);
-  const [appliedTagIds, setAppliedTagIds] = useState<string[]>([]);
-  const [appliedTagGroupKeys, setAppliedTagGroupKeys] = useState<string[]>([]);
-  const [appliedRegion, setAppliedRegion] = useState<string | null>(null);
   const [locating, setLocating] = useState(false);
   const [userLocation, setUserLocation] = useState<{ lat: number; lng: number } | null>(null);
   const { toast, showToast } = useToast();
   const mapRef = useRef<FocusCameraHandle>(null);
   const listScrollRef = useRef<HTMLDivElement | null>(null);
   const listScrollMemoRef = useRef<number>(0);
-  const urlFilterInitRef = useRef(false);
   const { recents, addRecent, removeRecent, clearRecents } = useRecentSearches();
   const { restored, save, clear } = useDiscoverViewState();
 
-  // allPlaces에서 등장하는 도시만 추출 (level=1은 parent로 rollup). useEffect보다 먼저 선언.
-  const availableCities = useMemo(() => {
-    const map = new Map<string, string>(); // slug → label(원본 nameEn)
-    for (const place of allPlaces) {
-      const slug = getPlaceRegionSlug(place.area);
-      const label = getPlaceRegionLabel(place.area);
-      if (!slug || !label) continue;
-      if (!map.has(slug)) map.set(slug, label);
-    }
-    return [...map.entries()]
-      .map(([slug, label]) => ({ slug, label }))
-      .sort((a, b) => a.label.localeCompare(b.label));
-  }, [allPlaces]);
+  const {
+    isFilterOpen,
+    stagedTopicIds,
+    stagedTagIds,
+    stagedRegion,
+    appliedTopicIds,
+    appliedTagIds,
+    appliedTagGroupKeys,
+    appliedRegion,
+    hasFilters,
+    hasPostLevelFilter,
+    availableCities,
+    topicChipMap,
+    tagChipMap,
+    tagGroupChipMap,
+    btsTopicId,
+    btsChipInfo,
+    commitFilters,
+    exitResultMode,
+    openFilter,
+    applyFilters,
+    closeFilter,
+    resetStaged,
+    removeAppliedTopic,
+    removeAppliedTag,
+    removeAppliedTagGroup,
+    toggleTopic,
+    toggleTag,
+    toggleRegion,
+  } = useDiscoverFilters({
+    topicTree,
+    tagGroups,
+    allPlaces,
+    onExitQuery: () => setQuery(""),
+    onFiltersApplied: () => setSheetState("half"),
+  });
 
   useEffect(() => {
     if (!restored) return;
@@ -145,24 +151,6 @@ export function ExploreMapView({ allPlaces, savedPostIds, savedEventIds = [], ta
       });
     }
   }, [restored, clear]);
-
-  // URL ?topics=<slug,…> / ?tags=<slug,…> / ?region=<slug> → 마운트 1회 초기 필터 적용
-  useEffect(() => {
-    if (urlFilterInitRef.current) return;
-    urlFilterInitRef.current = true;
-    const parsed = parseFilterParams(
-      new URLSearchParams(searchParams.toString()),
-      topicTree,
-      tagGroups
-    );
-    if (parsed.topicIds.length > 0) { setAppliedTopicIds(parsed.topicIds); setStagedTopicIds(parsed.topicIds); }
-    if (parsed.tagIds.length > 0) { setAppliedTagIds(parsed.tagIds); setStagedTagIds(parsed.tagIds); }
-    if (parsed.tagGroupKeys.length > 0) { setAppliedTagGroupKeys(parsed.tagGroupKeys); setStagedTagGroupKeys(parsed.tagGroupKeys); }
-    if (parsed.region && availableCities.some((c) => c.slug === parsed.region)) {
-      setAppliedRegion(parsed.region);
-      setStagedRegion(parsed.region);
-    }
-  }, [searchParams, topicTree, tagGroups, availableCities]);
 
   // 컬렉션 진입/변경/종료 시 이벤트 검색·카테고리·북마크 초기화
   useEffect(() => {
@@ -216,16 +204,6 @@ export function ExploreMapView({ allPlaces, savedPostIds, savedEventIds = [], ta
       params.delete("place");
     }
     router.replace(`?${params.toString()}`);
-  }
-
-  function commitFilters(next: { topicIds: string[]; tagIds: string[]; tagGroupKeys: string[]; region: string | null }) {
-    setAppliedTopicIds(next.topicIds);
-    setAppliedTagIds(next.tagIds);
-    setAppliedTagGroupKeys(next.tagGroupKeys);
-    setAppliedRegion(next.region);
-    const params = serializeFilterParams(next, { topicTree, tagGroups }, new URLSearchParams(searchParams.toString()));
-    const qs = params.toString();
-    router.replace(qs ? `${pathname}?${qs}` : pathname, { scroll: false });
   }
 
   const handleLocateMe = () => {
@@ -297,61 +275,6 @@ export function ExploreMapView({ allPlaces, savedPostIds, savedEventIds = [], ta
     [tagGroups]
   );
 
-  const topicChipMap = useMemo(() => {
-    const map = new Map<string, ChipInfo>();
-    for (const root of topicTree) {
-      const l1s = root.children;
-      if (l1s.length === 0) continue;
-      if (root.nameEn === KPOP_NAME) {
-        for (const l1 of l1s)
-          for (const l2 of l1.children) {
-            const r = resolveTopicColors({ ...l2, parent: { ...l1, parent: root } });
-            map.set(l2.id, { id: l2.id, label: l2.nameEn, bg: labelBackground({ text: "", ...r }), fg: r.textColorHex });
-          }
-        continue;
-      }
-      const hasL2 = l1s.some((l1) => l1.children.length > 0);
-      if (!hasL2) {
-        for (const l1 of l1s) {
-          const r = resolveTopicColors({ ...l1, parent: root });
-          map.set(l1.id, { id: l1.id, label: l1.nameEn, bg: labelBackground({ text: "", ...r }), fg: r.textColorHex });
-        }
-      } else {
-        for (const l1 of l1s)
-          for (const l2 of l1.children) {
-            const r = resolveTopicColors({ ...l2, parent: { ...l1, parent: root } });
-            map.set(l2.id, { id: l2.id, label: l2.nameEn, bg: labelBackground({ text: "", ...r }), fg: r.textColorHex });
-          }
-      }
-    }
-    return map;
-  }, [topicTree]);
-
-  const tagChipMap = useMemo(() => {
-    const map = new Map<string, ChipInfo>();
-    for (const group of tagGroups)
-      for (const tag of group.tags) {
-        const r = resolveTagColors(tag, group);
-        map.set(tag.id, { id: tag.id, label: tag.name, bg: labelBackground({ text: "", ...r }), fg: tag.textColorHex ?? group.textColorHex ?? DEFAULT_TEXT });
-      }
-    return map;
-  }, [tagGroups]);
-
-  const tagGroupChipMap = useMemo(() => {
-    const map = new Map<string, ChipInfo>();
-    for (const g of tagGroups) {
-      const bg = labelBackground({ text: "", colorHex: g.colorHex, colorHex2: g.colorHex2, gradientDir: g.gradientDir, gradientStop: g.gradientStop, textColorHex: g.textColorHex });
-      map.set(g.group, { id: g.group, label: g.nameEn, bg, fg: g.textColorHex });
-    }
-    return map;
-  }, [tagGroups]);
-
-  const btsTopicId = useMemo(() => topicSlugToId(topicTree, "bts"), [topicTree]);
-  const btsChipInfo = useMemo(
-    () => (btsTopicId ? (topicChipMap.get(btsTopicId) ?? null) : null),
-    [btsTopicId, topicChipMap]
-  );
-
   const markerPlaces = useMemo(
     () => allPlaces.map((p) => ({
       ...p,
@@ -376,8 +299,6 @@ export function ExploreMapView({ allPlaces, savedPostIds, savedEventIds = [], ta
   );
   const isEventMode = activeEventData !== null;
 
-  const hasFilters = appliedTopicIds.length > 0 || appliedTagIds.length > 0 || appliedTagGroupKeys.length > 0 || appliedRegion !== null;
-  const hasPostLevelFilter = appliedTopicIds.length > 0 || appliedTagIds.length > 0 || appliedTagGroupKeys.length > 0;
   const isResultMode = !isEventMode && (query.trim() !== "" || hasFilters);
   const hasRegionChips = !isEventMode && availableCities.length >= 2;
   // 이벤트 모드는 EventSearchBar(검색+칩)가 항상 떠 있어 동일한 top reserve가 필요.
@@ -648,38 +569,6 @@ export function ExploreMapView({ allPlaces, savedPostIds, savedEventIds = [], ta
     fitEventMarkers({ query: eventQuery, category: selectedCategory, savedOnly: next, savedSet: savedEventIdsSet });
   }
 
-  const exitResultMode = () => {
-    setQuery("");
-    commitFilters({ topicIds: [], tagIds: [], tagGroupKeys: [], region: null });
-  };
-
-  const openFilter = () => {
-    setStagedTopicIds(appliedTopicIds);
-    setStagedTagIds(appliedTagIds);
-    setStagedTagGroupKeys(appliedTagGroupKeys);
-    setStagedRegion(appliedRegion);
-    setIsFilterOpen(true);
-  };
-  const applyFilters = () => {
-    commitFilters({ topicIds: stagedTopicIds, tagIds: stagedTagIds, tagGroupKeys: stagedTagGroupKeys, region: stagedRegion });
-    setIsFilterOpen(false);
-    setSheetState("half");
-  };
-  const resetStaged = () => { setStagedTopicIds([]); setStagedTagIds([]); setStagedTagGroupKeys([]); setStagedRegion(null); };
-  const removeAppliedTopic = (id: string) =>
-    commitFilters({ topicIds: appliedTopicIds.filter((x) => x !== id), tagIds: appliedTagIds, tagGroupKeys: appliedTagGroupKeys, region: appliedRegion });
-  const removeAppliedTag = (id: string) =>
-    commitFilters({ topicIds: appliedTopicIds, tagIds: appliedTagIds.filter((x) => x !== id), tagGroupKeys: appliedTagGroupKeys, region: appliedRegion });
-  const removeAppliedTagGroup = (key: string) =>
-    commitFilters({ topicIds: appliedTopicIds, tagIds: appliedTagIds, tagGroupKeys: appliedTagGroupKeys.filter((k) => k !== key), region: appliedRegion });
-
-  const toggleTopic = (id: string) =>
-    setStagedTopicIds((prev) => prev.includes(id) ? prev.filter((x) => x !== id) : [...prev, id]);
-  const toggleTag = (id: string) =>
-    setStagedTagIds((prev) => prev.includes(id) ? prev.filter((x) => x !== id) : [...prev, id]);
-  const toggleRegion = (slug: string) =>
-    setStagedRegion((prev) => prev === slug ? null : slug);
-
   const effectiveSheetState = selectedPlaceId
     ? "hidden"
     : sheetState === "hidden"
@@ -924,7 +813,7 @@ export function ExploreMapView({ allPlaces, savedPostIds, savedEventIds = [], ta
       {/* 필터 시트 — z-[65] */}
       <DiscoverFilterSheet
         isOpen={isFilterOpen}
-        onClose={() => setIsFilterOpen(false)}
+        onClose={closeFilter}
         topicTree={topicTree}
         tagGroups={tagGroups}
         topicChipMap={topicChipMap}
